done add image to show data

+++ chp6/admin.php
@@ -27,7 +27,16 @@ while($row = mysqli_fetch_array($data)){
     echo '<tr class="scorerow"><td><strong>'.$row['name'].'</strong></td>';
     echo '<td>'.$row['date'].'</td>';
     echo '<td>'.$row['score'].'</td>';
-    echo '<td>'.$row['screenshot'].'</td>';
+
+        if(is_file(GW_UPLOADPATH . $row['screenshot']) && filesize(GW_UPLOADPATH . $row['screenshot'])>0){
+            echo '<td><img src="' .GW_UPLOADPATH. $row['screenshot'] . '" alt="Score image"  width="250"/></td>';
+        }
+        else{
+
+            echo '<td>'.$row['screenshot'].'</td>';
+        }
+
+
     echo '<td><a 
 href="removescore.php?id='.$row['id'].'&amp;date='.$row['date'].
         '&amp;name='.$row['name'].'&amp;score='.$row['score'].'&amp;screenshot='.$row['screenshot'].'">Remove</a> ';
